Implement NodeRule trait for Midpoint and Simpson
